rename the create_tag to just create
